% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesrep-package.R
\docType{package}
\name{bayesrep-package}
\alias{bayesrep}
\alias{bayesrep-package}
\title{bayesrep: Bayesian Repulsive Ensembles via Ising Priors}
\description{
Fits ensembles of sparse linear and logistic sub-models whose feature
allocations repel one another under an Ising prior, so that collinear
predictive pathways are separated across the ensemble instead of being
arbitrarily collapsed onto a single representative.
}
\details{
The entry point is \code{\link{bre}}. See \code{\link{coallocation}} for the
label-invariant allocation summary and \code{\link{waic}} for choosing the
number of sub-models.
}
\author{
\strong{Maintainer}: Anthony Christidis \email{anthony.christidis@stat.ubc.ca}

}
\keyword{internal}
