% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvBre.R
\name{plot.cv.bre}
\alias{plot.cv.bre}
\title{Plot a cross-validation curve}
\usage{
\method{plot}{cv.bre}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cv.bre"}.}

\item{...}{Unused, for method consistency.}
}
\value{
Invisibly \code{NULL}; called for the plot side effect.
}
\description{
Plot a cross-validation curve
}
