% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.bre}
\alias{print.bre}
\title{Print a Bayesian Repulsive Ensemble fit}
\usage{
\method{print}{bre}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"bre"}.}

\item{...}{Unused, for method consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Bayesian Repulsive Ensemble fit
}
