% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvBre.R
\name{print.cv.bre}
\alias{print.cv.bre}
\title{Print a cross-validated BRE selection}
\usage{
\method{print}{cv.bre}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"cv.bre"}.}

\item{...}{Unused, for method consistency.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a cross-validated BRE selection
}
